API: configure Cython build via environment variables
Previously configuration via `pip` used `--install-option`,
but this parameter is removed in `pip >= 23.1`.

- REL: use `pip`, instead of `setup.py install`,
  which is deprecated in `setuptools == 58.3.0`.

- REL: require `setuptools >= 65.6.0` in `install_requires`,
  to use `pip install . --use-pep517`

- CI: use `pip` to install `dd`

- API: warn about `extern/` license files when running
  `bdist_wheel`, to enable `pip install .` for local builds.

+++ download.py
@@ -174,8 +174,12 @@ def _copy_extern_licenses(args):
     for name in licenses:
         license = os.path.join(path, name)
         included = os.path.join('dd', name)
-        if yes:
+        if yes and os.path.isfile(license):
             shutil.copyfile(license, included)
+        elif yes and not os.path.isfile(license):
+            print(
+                f'WARNING: No file: `{license}`, '
+                'skipping file copy.')
         elif os.path.isfile(included):
             os.remove(included)
